src/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
asyncpg>=0.29,<1.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
email-validator>=2.2,<3.0
python-multipart>=0.0.9,<1.0
python-dotenv>=1.0,<2.0
bcrypt=4.0.1
